+++ src/main/java/sol3675/middleearththaumaturgy/common/items/wands/FactionWandUpdateTT.java
@@ -1,6 +1,6 @@
 package sol3675.middleearththaumaturgy.common.items.wands;
 
-import lotr.common.LOTRFaction;
+import lotr.common.fac.LOTRFaction;
 import lotr.common.LOTRLevelData;
 import lotr.common.LOTRPlayerData;
 import net.minecraft.entity.player.EntityPlayer;
@@ -13,22 +13,22 @@ public class FactionWandUpdateTT extends FactionWandUpdate{
 	}
 	
 	@Override
-	int getVisRecovery(int alignment)
+	int getVisRecovery(float alignment)
 	{
 		int i = 0;
-		if (alignment >= 100000)
+		if (alignment >= 100000.0F)
 		{
 			i = 400;
 		}
-		else if (alignment >= 10000)
+		else if (alignment >= 10000.0F)
 		{
 			i = 100;
 		}
-		else if (alignment >= 1000)
+		else if (alignment >= 1000.0F)
 		{
 			i = 20;
 		}
-		else if (alignment >= 100)
+		else if (alignment >= 100.0F)
 		{
 			i = 10;
 		}
